Relocated Autolink functions.
Removed from imgboard.
Put in _core/thread/autolink.php.

These could potentially be replicated in Regist's text processors.
